+++ eslint.config.js
@@ -43,6 +43,7 @@ const rules = {
   'no-undef': 2,
   'no-underscore-dangle': 0,
   'no-unused-expressions': 2,
+  'no-unused-vars': 2,
   'no-use-before-define': 0,
   'no-var': 2,
   'no-with': 2,