+++ test/helpers/setup.js
@@ -1,5 +1,0 @@
-import 'chai/register-expect.js';
-
-import nock from 'nock';
-
-nock.enableNetConnect(/(localhost|127\.0\.0\.1):\d+/);